+++ src/gConsent/manage/approveAccessRequest.ts
@@ -51,7 +51,7 @@ import { getBaseAccess } from "../util/getBaseAccessVerifiableCredential";
 import { initializeGrantParameters } from "../util/initializeGrantParameters";
 import { getGrantBody, issueAccessVc } from "../util/issueAccessVc";
 import { toVcDataset } from "../../common/util/toVcDataset";
-import type { CustomFields } from "../type/CustomFields";
+import type { CustomFields } from "../../type/CustomFields";
 
 export type ApproveAccessRequestOverrides = Omit<
   Omit<AccessGrantParameters, "status">,
